+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - **Pytest**: Cleaner error reporting for Schema-related errors (like unknown media type).
 
+### :bug: Fixed
+
+- **Pytest**: Issue where enabling or disabling checks via the config file had no effect.
+
 ## [4.0.2](https://github.com/schemathesis/schemathesis/compare/v4.0.1...v4.0.2) - 2025-06-15
 
 ### :bug: Fixed