+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### :bug: Fixed
 
 - **Pytest**: Issue where enabling or disabling checks via the config file had no effect.
+- **Pytest**: Ignored `tls-verify`, `request-timeout`, `request-cert`, and `request-cert-key` options supplied via the config file.
 
 ## [4.0.2](https://github.com/schemathesis/schemathesis/compare/v4.0.1...v4.0.2) - 2025-06-15